osa_expert/pyproject.toml
[project]
name = "osa-expert"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ainish-coder bot", email = "bot@local" }
]
requires-python = ">=3.10"
dependencies = [
    "dspy-ai>=3.1.0",
    "lightrag-hku>=1.4.9.10",
    "mcp>=1.25.0",
    "memori>=3.1.3",
    "raganything[all]>=1.2.8",
]

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"
